Refactor coordinate types to double
This commit updates several classes and interfaces to support more precise coordinate values by changing the type of coordinates from `int` to `double`. Specifically, changes were made in `TwoDimension.cs`, `IPoint.cs`, `Point.cs`, and `Point2D.cs` to ensure consistency and flexibility in handling points with non-integer coordinates. The `GetClosestPointsSplit` method in `TwoDimension.cs`, and the `GetCoordinate` method in the `IPoint` interface, along with the constructor parameters and `X` (and `Y` where applicable) fields in `Point.cs` and `Point2D.cs`, have been updated to reflect this change. This modification allows for a more accurate representation of points and their coordinates.

+++ Spacing/Spacing.Common/TwoDimension.cs
@@ -98,7 +98,7 @@ public class TwoDimension
     private static (Point2D x, Point2D y, double d) GetClosestPointsSplit(Point2D[] pointsX, Point2D[] pointsY, double currentClosestDistance)
     {
         // the X coordinate of the last point in array is median because array is sorted and half divided
-        int medianX = pointsX[^1].GetCoordinate(1);
+        double medianX = pointsX[^1].GetCoordinate(1);
         // the range of X coordinates to check
         double xMin = medianX - currentClosestDistance;
         // the range of X coordinates to check
